rename time_manager file to search_limiter
